organize torus-specific solver
